# docs/openapi/v2.yaml
openapi: 3.1.0
info:
  title: Bazaar v2 API
  version: 2.0.0
  description: |
    Unified BFF API for the Bazaar platform.  All endpoints live under `/v2/*`.

    **Platform routing** — every request must carry an `X-Client` header:
    - `mobile` — Flutter app (bearer tokens).
    - `web` — Next.js storefront (cookies preferred).

    Endpoints marked **mobile-only** or **web-only** return `400 UNKNOWN_PLATFORM`
    when called from the wrong client.  Endpoints under the **Shared** group
    (Products, Wishlist) are served to both clients.

    **Response envelope** — every response, success or error, is wrapped:

    ```json
    { "success": true, "data": { … } }
    { "success": false, "error": { "code": "TOKEN_EXPIRED", "message": "…" } }
    ```

    Paginated responses add a `meta` key alongside `data`.

    **Auth** — two complementary schemes:
    - `bearerAuth` — `Authorization: Bearer <access-jwt>` (mobile default, web fallback).
    - `cookieAuth` — `bz_at` HttpOnly cookie (web default).

    The app accepts either; prefer what matches `tokenDelivery` at login.

    **Error codes** — machine-readable codes in `error.code`.  Full catalog:
    `docs/v2-auth/ERROR-CATALOG.md`.
  contact:
    name: Bazaar Platform Team
    email: platform@bazaar-uae.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: "${BACKEND_URL}/v2"
    description: Production (set BACKEND_URL in .env)
  - url: "${STAGING_URL}/v2"
...

...
# Tags
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access JWT obtained from `/v2/auth/login` (or OAuth login) with
        `tokenDelivery: "bearer"`.  Pass as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: bz_at
      description: >
        HttpOnly access-token cookie set by the server when
        `tokenDelivery: "cookie"` is used.  Web clients use this by default.

# ─────────────────────────────────────────────────────────────────────────────
# Reusable Schemas
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [Auth]
      summary: Register a new account
      description: |
        Creates a new user account and returns a token pair.
        Available on **both** mobile and web clients.
        Rate limit: 5 requests / IP / hour.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSuccess'
        '409':
          $ref: '#/components/responses/409Conflict'
        '422':
          $ref: '#/components/responses/422Unprocessable'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/login:
    post:
      operationId: authLogin
      tags: [Auth]
      summary: Email + password login
      description: |
        Authenticates with email and password.  Returns a token pair (status `ok`)
        or an MFA challenge (status `mfa_required`).
        Available on **both** mobile and web clients.
        Rate limit: 10 / account / 15 min.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful or MFA challenge issued.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/AuthSuccess'
                          - $ref: '#/components/schemas/MfaChallenge'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '422':
          $ref: '#/components/responses/422Unprocessable'
        '423':
          description: Account locked after too many failed attempts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/login/google:
    post:
      operationId: authLoginGoogle
      tags: [Auth]
      summary: Google OAuth login / signup
      description: |
        Authenticates or registers via a Google id-token.
        Available on **both** mobile and web clients.
        Rate limit: 10 / IP / minute.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthLoginRequest'
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSuccess'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          $ref: '#/components/responses/409Conflict'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/login/apple:
    post:
      operationId: authLoginApple
      tags: [Auth]
      summary: Apple OAuth login / signup
      description: |
        Authenticates or registers via an Apple id-token.
        Available on **both** mobile and web clients.
        Rate limit: 10 / IP / minute.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/OAuthLoginRequest'
                - type: object
                  properties:
                    authorizationCode:
                      type: string
                      nullable: true
                    name:
                      type: object
                      nullable: true
                      properties:
                        firstName:
                          type: string
                          nullable: true
                        lastName:
                          type: string
                          nullable: true
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSuccess'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          $ref: '#/components/responses/409Conflict'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/password/forgot:
    post:
      operationId: authPasswordForgot
      tags: [Auth]
      summary: Request password reset email
      description: |
        Sends a 6-digit recovery code to the email address.
        Always returns success to prevent email enumeration.
        Available on **both** mobile and web clients.
        Rate limit: 3 / email / hour.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Email sent (or silently accepted if email not found).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          sent:
                            type: boolean
                            example: true
        '422':
          $ref: '#/components/responses/422Unprocessable'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/password/verify-code:
    post:
      operationId: authPasswordVerifyCode
      tags: [Auth]
      summary: Exchange recovery code for reset token
      description: |
        Validates the 6-digit emailed code and returns a short-lived reset token.
        Available on **both** mobile and web clients.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: Code verified — reset token issued.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          resetToken:
                            type: string
                          expiresIn:
                            type: integer
                            example: 900
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/password/reset:
    post:
      operationId: authPasswordReset
      tags: [Auth]
      summary: Set new password using reset token
      description: |
        Sets a new password using the reset token from `/auth/verify-code`.
        Revokes all active sessions. User must log in again after reset.
        Available on **both** mobile and web clients.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          passwordReset:
                            type: boolean
                            example: true
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422Unprocessable'
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/recovery/verify:
    post:
      operationId: authVerifyRecovery
      tags: [Auth]
      summary: Verify account recovery code
      description: |
        Verifies a recovery code for account recovery flow.
        Available on **both** mobile and web clients.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryAccountRequest'
      responses:
        '200':
          description: Recovery code verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '422':
          $ref: '#/components/responses/422Unprocessable'
        '500':
          $ref: '#/components/responses/500Internal'
  /auth/recovery/resend:
    post:
      operationId: authResendRecovery
      tags: [Auth]
      summary: Resend account recovery code
      description: |
        Resends the account recovery code email.
        Available on **both** mobile and web clients.
        Rate limit: 3 / account / 24 hours.
      security: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendRecoveryRequest'
      responses:
        '200':
          description: Recovery code resent.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          sent:
                            type: boolean
                            example: true
        '429':
          $ref: '#/components/responses/429RateLimited'
        '500':
          $ref: '#/components/responses/500Internal'
...
  /cart:
    get:
      operationId: cartGet
      tags: [Cart]
      summary: Get the current user's cart
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/XClient'
      responses:
        '200':
          description: Cart contents.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '500':
          $ref: '#/components/responses/500Internal'
...